+++ src/Symfony/Component/Messenger/Tests/Transport/Serialization/SerializerTest.php
@@ -14,7 +14,6 @@ namespace Symfony\Component\Messenger\Tests\Transport\Serialization;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Exception\MessageDecodingFailedException;
-use Symfony\Component\Messenger\Stamp\NonSendableStampInterface;
 use Symfony\Component\Messenger\Stamp\SerializedMessageStamp;
 use Symfony\Component\Messenger\Stamp\SerializerStamp;
 use Symfony\Component\Messenger\Stamp\ValidationStamp;
@@ -208,18 +207,6 @@ class SerializerTest extends TestCase
         ]);
     }
 
-    public function testEncodedSkipsNonEncodeableStamps()
-    {
-        $serializer = new Serializer();
-
-        $envelope = new Envelope(new DummyMessage('Hello'), [
-            new DummySymfonySerializerNonSendableStamp(),
-        ]);
-
-        $encoded = $serializer->encode($envelope);
-        $this->assertStringNotContainsString('DummySymfonySerializerNonSendableStamp', print_r($encoded['headers'], true));
-    }
-
     public function testDecodingFailedConstructorDeserialization()
     {
         $serializer = new Serializer();
@@ -246,9 +233,6 @@ class SerializerTest extends TestCase
             ],
         ]);
     }
-}
-class DummySymfonySerializerNonSendableStamp implements NonSendableStampInterface
-{
 }
 class DummySymfonySerializerInvalidConstructor
 {