+++ src/Symfony/Component/Messenger/Transport/Serialization/PhpSerializer.php
@@ -13,7 +13,6 @@ namespace Symfony\Component\Messenger\Transport\Serialization;
 
 use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Exception\MessageDecodingFailedException;
-use Symfony\Component\Messenger\Stamp\NonSendableStampInterface;
 
 /**
  * @author Ryan Weaver<ryan@symfonycasts.com>
@@ -43,8 +42,6 @@ class PhpSerializer implements SerializerInterface
      */
     public function encode(Envelope $envelope): array
     {
-        $envelope = $envelope->withoutStampsOfType(NonSendableStampInterface::class);
-
         $body = addslashes(serialize($envelope));
 
         if (!preg_match('//u', $body)) {